+++ generators/app/templates/api/controllers/SearchController.js
@@ -1,0 +1,11 @@
+/**
+ * SearchController
+ * @description :: Server-side logic for searching
+ */
+
+module.exports = {
+  index: function (req, res) {
+    //   TODO: implement search
+    res.badRequest(null, null, 'Not implemented yet...');
+  }
+};