+++ generators/app/templates/config/http.js
@@ -4,51 +4,74 @@
  * Configuration for the underlying HTTP server in Sails
  */
 
-module.exports.http = {
+module.exports = {
   /**
-   * This is the options object for the `createServer` method, as discussed here:
-   * http://nodejs.org/api/https.html#https_class_https_server
-   *
-   * @type {Object|Boolean}
+   * Port where to run this app
+   * @type {Number}
    */
-  serverOptions: undefined,
+  port: 3000,
 
   /**
-   * You can define own custom middleware here
-   * @param app Express application
+   * Your SSL certificates
+   * @description They are should be as result of fs.readFileSync() method
+   * @type {Object}
+   * @example
+   * ssl: {
+   *    cert: fs.readFileSync('your/path/to/cert.pem'),
+   *    key: fs.readFileSync('your/path/to/key.pem')
+   * }
    */
-  customMiddleware: function (app) {
-
+  ssl: {
+    cert: false,
+    key: false
   },
 
-  /**
-   * Configures the middleware function used for parsing the HTTP request body
-   * Defaults to the Formidable-based version built into Express/Connect
-   *
-   * To enable streaming file uploads (to disk or somewhere else)
-   * you'll want to set this option to `false` to disable the body parser
-   *
-   * @type {Function|Boolean|Object}
-   */
-  bodyParser: undefined,
-
-  /**
-   * Express middleware to use for every Sails request
-   * @type {Object}
-   */
-  middleware: {
+  http: {
     /**
-     * The order in which middleware should be run for HTTP request
-     * @type {Array}
+     * This is the options object for the `createServer` method, as discussed here:
+     * http://nodejs.org/api/https.html#https_class_https_server
+     *
+     * @type {Object|Boolean}
      */
-    order: [
-      'bodyParser',
-      'compress',
-      'methodOverride',
-      '$custom',
-      'router',
-      '404',
-      '500'
-    ]
+    serverOptions: undefined,
+
+    /**
+     * You can define own custom middleware here
+     * @param app Express application
+     */
+    customMiddleware: function (app) {
+
+    },
+
+    /**
+     * Configures the middleware function used for parsing the HTTP request body
+     * Defaults to the Formidable-based version built into Express/Connect
+     *
+     * To enable streaming file uploads (to disk or somewhere else)
+     * you'll want to set this option to `false` to disable the body parser
+     *
+     * @type {Function|Boolean|Object}
+     */
+    bodyParser: undefined,
+
+    /**
+     * Express middleware to use for every Sails request
+     * @type {Object}
+     */
+    middleware: {
+      /**
+       * The order in which middleware should be run for HTTP request
+       * @type {Array}
+       */
+      order: [
+        'bodyParser',
+        'compress',
+        'methodOverride',
+        '$custom',
+        'router',
+        '404',
+        '500'
+      ]
+    }
   }
 };