+++ generators/app/templates/api/services/CipherService.js
@@ -1,4 +1,1 @@
-var CipherFactory = require('./cipher/CipherFactory');
-
-module.exports = new CipherFactory().create("<%= answers['services:cipher'].toLowerCase() %>");
-module.exports.Factory = new CipherFactory();
+module.exports = require('sails-service-cipher');