+++ generators/app/templates/api/services/MailerService.js
@@ -1,4 +1,1 @@
-var MailerFactory = require('./mailer/MailerFactory');
-
-module.exports = new MailerFactory().create("<%= answers['services:mailer'].toLowerCase() %>");
-module.exports.Factory = new MailerFactory();
+module.exports = require('sails-service-mailer');