+++ generators/app/templates/api/services/PaymentService.js
@@ -1,4 +1,1 @@
-var PaymentFactory = require('./payment/PaymentFactory');
-
-module.exports = new PaymentFactory().create("<%= answers['services:payment'].toLowerCase() %>");
-module.exports.Factory = new PaymentFactory();
+module.exports = require('sails-service-payment');