+++ generators/app/templates/api/services/PusherService.js
@@ -1,4 +1,1 @@
-var PusherFactory = require('./pusher/PusherFactory');
-
-module.exports = new PusherFactory();
-module.exports.Factory = new PusherFactory();
+module.exports = require('sails-service-pusher');