+++ generators/app/templates/api/services/SmsService.js
@@ -1,4 +1,1 @@
-var SmsFactory = require('./sms/SmsFactory');
-
-module.exports = new SmsFactory().create("<%= answers['services:sms'].toLowerCase() %>");
-module.exports.Factory = new SmsFactory();
+module.exports = require('sails-service-sms');