+++ generators/app/templates/api/services/SocialService.js
@@ -1,4 +1,1 @@
-var SocialFactory = require('./social/SocialFactory');
-
-module.exports = new SocialFactory();
-module.exports.Factory = new SocialFactory();
+module.exports = require('sails-service-social');