+++ generators/app/templates/api/services/StorageService.js
@@ -1,4 +1,1 @@
-var StorageFactory = require('./storage/StorageFactory');
-
-module.exports = new StorageFactory().create("<%= answers['services:storage'].toLowerCase() %>");
-module.exports.Factory = new StorageFactory();
+module.exports = require('sails-service-storage');