+++ generators/app/templates/config/cron.js
@@ -1,0 +1,5 @@
+var CronJob = require('cron').CronJob;
+
+new CronJob('* * * * * *', function onTick() {
+  //TODO: you can fill it with your job
+}, null, true);