+++ generators/app/templates/api/blueprints/findone.js
@@ -1,3 +1,4 @@
+var _ = require('lodash');
 var actionUtil = require('sails/lib/hooks/blueprints/actionUtil');
 
 /**
@@ -7,7 +8,14 @@ var actionUtil = require('sails/lib/hooks/blueprints/actionUtil');
  * An API call to find and return a single model instance from the data adapter using the specified id.
  */
 module.exports = function (req, res) {
-  actionUtil.populateEach(actionUtil.parseModel(req).findOne(actionUtil.requirePk(req)), req)
+  _.set(req.options, 'criteria.blacklist', ['limit', 'skip', 'sort', 'populate', 'fields']);
+
+  var fields = req.param('fields') ? req.param('fields').split(',') : false;
+  var Model = actionUtil.parseModel(req);
+  var PK = actionUtil.requirePk(req);
+  var findQuery = actionUtil.populateEach(Model.findOne(PK), req);
+
+  findQuery
     .then(function (record) {
       if (!record) return res.notFound();