+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Improvement: Remove old `isAllowed` policy by `Application-Token`. Will be more secured variant;
 - Improvement: Change JWT algorithm to HS512;
+- Improvement: Replace `jwt.decode()` with `jwt.verify()`. It's security fix, because decode doesn't verify token;
+- Improvement: Allow to send options object to `jsonwebtoken` verify and sign methods;
 
 ## Version 0.9.2