+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Typo: Change JWT algorithm to HS512;
 - Typo: Remove old `isAllowed` policy by `Application-Token`;
 - Typo: Remove `skip-welcome` option;
+- Typo: Rename `skip-update` option to `skip-check-update`;
 
 ## Version 0.9.2