+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Edge version
 
+## Version 0.10.0
+
 - **Improvement**: I have moved all the services from generator to separate npm modules. In this way you still can get updates for services and get hotfixes and new features without regenerating the whole project;
 - Improvement: Increase stability level index. So we are moving to stable versions;
 - Improvement: Implement `fields` parameter in requests that allows you set name of fields that you want to get;